unify native Python operator API
