+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/common/operators/source/PythonSourceOperatorDescriptor.scala
@@ -1,0 +1,7 @@
+package edu.uci.ics.texera.workflow.common.operators.source
+
+import edu.uci.ics.texera.workflow.common.operators.PythonOperatorDescriptor
+
+abstract class PythonSourceOperatorDescriptor extends SourceOperatorDescriptor with PythonOperatorDescriptor {
+
+}
